Se agregaron alertas, al dar de baja, alta o modificar elementos. Se agregó un eliminado multiple

+++ src/views/TablaConBotones.js
@@ -1,7 +1,17 @@
 // ** React Imports
 import { Fragment, useState, forwardRef, useEffect, useRef } from 'react'
 
-import { Link,  withRouter } from 'react-router-dom'
+import Swal from "sweetalert2"
+
+const customSwal = Swal.mixin({
+  customClass: {
+    confirmButton: 'btn btn-primary mx-1',
+    cancelButton: 'btn btn-outline-primary mx-1'
+  },
+  buttonsStyling: false
+})
+
+import { Link } from 'react-router-dom'
 
 import { useSelector, useDispatch } from 'react-redux'
 import devActions from '../redux/actions/table/devActions'
@@ -18,7 +28,8 @@ import {
   UncontrolledDropdown,
   DropdownToggle,
   DropdownMenu,
-  DropdownItem
+  DropdownItem,
+  Button
 } from 'reactstrap'
 
 // ** Bootstrap Checkbox Component
@@ -29,11 +40,19 @@ const BootstrapCheckbox = forwardRef(({ onClick, ...rest }, ref) => (
   </div>
 ))
 
-const TablaConBotones = (props) => {
+const TablaConBotones = () => {
   const URL = process.env.REACT_APP_API_URL
   const axios = require('axios')
   const devs = useSelector(state => state.devs)
   const dispatch = useDispatch()
+
+  const [checked, setChecked] = useState(true)
+  const [selectedRows, setSelectedRows] = useState([])
+
+  const handleChange = (state) => {
+    setSelectedRows(state.selectedRows)
+    setChecked(true)
+  }
 
   const loadData = () => {
     dispatch(devActions.fetchDevsPending())
@@ -51,10 +70,51 @@ const TablaConBotones = (props) => {
         console.log(error)
       })
   }
-  const deleteDev = async (id) => {
-    console.log('dentro de la funcion')
-    const res = await axios.delete(`${URL}/${id}`)
-    loadData()
+  const deleteDev = (id) => {
+    customSwal.fire({
+      title: "¿Eliminar?",
+      text: "Esta accion no se puede revertir",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar"
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const res = await axios.delete(`${URL}/${id}`)
+        if (res.status !== 200) {
+          customSwal.fire("Error", "Espere unos minutos", "error")
+        }
+      }
+    }).then(() => loadData())
+
+  }
+
+  const multipleDelete = (array) => {
+    customSwal.fire({
+      title: "¿Eliminar multiple?",
+      text: "Esta accion no se puede revertir",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar"
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        for (const element of array) {
+          const res = await axios.delete(`${URL}/${element.id}`)
+          if (res.status !== 200) {
+            customSwal.fire("Error", "Espere unos minutos", "error")
+          }
+        }
+      }
+    }).then(() => {
+      loadData()
+      setSelectedRows([])
+      setChecked(false)
+    })
 
   }
 
@@ -91,11 +151,11 @@ const TablaConBotones = (props) => {
                 <MoreVertical size={15} />
               </DropdownToggle>
               <DropdownMenu right>
-                <DropdownItem>
+                <Link to={`/inicio:table/editar/${row.id}`} className='dropdown-item'>
                   <Edit size={15} />
-                  <span onClick={() => props.history.push(`/inicio:table/editar/${row.id}`) } className='align-middle ml-50'>Editar</span>
-                </DropdownItem>
-                <DropdownItem>
+                  <span className='align-middle ml-50 '>Editar</span>
+                </Link>
+                <DropdownItem className='w-100'>
                   <Trash size={15} />
                   <span onClick={() => deleteDev(row.id)} className='align-middle ml-50'>Borrar</span>
                 </DropdownItem>
@@ -162,19 +222,18 @@ const TablaConBotones = (props) => {
         <CardHeader className='flex-md-row flex-column align-items-center border-bottom'>
           <CardTitle tag='h4'>Tabla de desarrolladores</CardTitle>
           <div className='d-flex mt-md-0 mt-1'>
-            <Link to='/inicio:table/agregar' className='ml-2 btn btn-primary'>
+            <Link to='/inicio:table/agregar' className='btn btn-primary'>
               <Plus size={15} />
               <span className='align-middle ml-50'>Agregar</span>
             </Link>
           </div>
         </CardHeader>
-        <CardBody className='dataTable'>
+        <CardBody className={`dataTable ${selectedRows.length > 0 ? 'grow' : ''}`}>
           <DataTable
             noHeader
             pagination
             selectableRows
             columns={columnas}
-            paginationPerPage={7}
             className='react-dataTable'
             sortIcon={<ChevronDown size={10} />}
             paginationDefaultPage={currentPage + 1}
@@ -184,11 +243,17 @@ const TablaConBotones = (props) => {
             selectableRowsComponent={BootstrapCheckbox}
             theme="solarized"
             customStyles={customStyles}
+            onSelectedRowsChange={handleChange}
+            clearSelectedRows={checked}
           />
+          <div className={`d-flex btnEliminarWrapper ${checked ? '' : 'd-none'}`}>
+            <Button onClick={() => multipleDelete(selectedRows)} className={`btnEliminar animate__animated  ${selectedRows.length > 0 ? 'animate__fadeInDown' : 'animate__fadeOutUp'}`} outline color='primary'><Trash size={15} /><span className='align-middle ml-50'>Eliminar</span></Button>
+          </div>
+
         </CardBody>
       </Card>
     </Fragment>
   )
 }
 
-export default withRouter(TablaConBotones)
+export default TablaConBotones
